IOS-7108: Perform network response mapping using dedicated mapper, improve model structure
